==> R/scripts/1_pseudoabsence_generation.Rmd <==
---
title: "Pseudoabsence generation"
author: "UK Centre for Ecology and Hydrology - DECIDE"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE) # always echo
```

## Intro

This script generates pseudoabsence (PAs) records to use in fitting statistical models for the SDMs.

These PAs are generated via phenologically aware targetting background approach.

```{r setup}
#file.path()

# Simon computer
#setwd("~/R/DECIDE_SDMs")

#JASMIN
#setwd("~/R/DECIDE_SDMs")

#Datalabs


library(readr)
library(dplyr)
library(tidyr)
library(ggplot2)
library(rJava)


library(parallel)
n.cores <- detectCores()
n.cores

source("R/functions/cpa.R") #function for generating absences


```

## Preparing phenology data

```{r phenology}
#load in the ecological traits data
#https://catalogue.ceh.ac.uk/documents/5b5a13b6-2304-47e3-9c9d-35237d1232c6
ecological_traits <- read_csv("data/raw_data/traits/butterfly_moth_ecological_traits.csv", 
    skip = 1)

#create phenology data in long format
phenology <- ecological_traits %>% dplyr::select(
  scientific_name,
  jan = ad_jan,
  feb = ad_feb,
  mar = ad_mar,
  apr = ad_apr,
  may = ad_may,
  jun = ad_jun,
  jul = ad_jul,
  aug = ad_agu, # note renaming here to be inline with tolower(month.abb)
  sep = ad_sep,
  oct = ad_oct,
  nov = ad_nov,
  dec = ad_dec) %>% 
  pivot_longer(-scientific_name,names_to = "month") %>%
  mutate(value = if_else(value == 1,T,F,missing = F))

phenology <- phenology %>% mutate(month_num = match(month, tolower(month.abb)))

saveRDS(phenology,"data/derived_data/species/phenology.RDS")


#phenology %>% filter(scientific_name == "Paranthrene tabaniformis")


```

## Generating PAs

Generate some test data. This is used for smaller scale testing. 

```{r test1}
if(F){
  #generate pseudo absences
  test_species <- distinct(phenology$scientific_name)[1:50]
  n_records <- 500000
  
  #species data for all species for all year round
  all_spdata <- data.frame(
    id = 1:n_records,
    species = sample(test_species,n_records,T),
    lat = sample(1:100,n_records,T),
    lon = sample(1:100,n_records,T),
    date = sample(seq(as.Date('2001/01/01'), as.Date('2020/01/01'), by="day"), n_records,T),
    group = rep("butterfly",n_records)
    )
  
  #get month as a number (1-12)
  all_spdata <- all_spdata %>% mutate(month = as.numeric(as.factor(months(date))))
  
  #str(all_spdata)
  #cpa(spdat,test_species[1],100,F,5,T)
  #species trends through the year
  # all_spdata %>% 
  #   group_by(species,month) %>%
  #   summarise(n = n()) %>%
  #   ggplot(aes(x = month,y = n,colour =species))+geom_line()+theme_bw()
}
```

Load real data.

```{r load_data}
#load in the data (pre-processed to remove duplicates etc.)
butterfly_records <- read_csv("data/derived_data/species/butterfly/records/butterfly_EastNorths_no_duplicates_2021_12_06.csv")
day_moth_records <- read_csv("data/derived_data/species/day_flying_moth/records/DayFlyingMoths_EastNorths_no_duplicates.csv")
night_moth_records <- read_csv("data/derived_data/species/day_flying_moth/records/NightFlyingMoths_EastNorths_no_duplicates_2021_12_06.csv")

# add a new column to differentiate the groups
butterfly_records$group <- "butterfly"
day_moth_records$group <- "day_flying_moth"
night_moth_records$group <- "night_flying_moth"

#make into one big dataframe
#select relevant columns ready for cpa function
all_spdata <- bind_rows(butterfly_records,day_moth_records) %>% 
  select(id = TO_ID,
         species = sp_n,
         lat = lat,
         lon = lon,
         date = date,
         group = group) %>%
  mutate(month = as.numeric(as.factor(months(date))))


species_list <- all_spdata %>% select(group,species) %>% distinct()
saveRDS(species_list,"data/derived_data/species/species_list.RDS")

#add a species group colomn to the phenology data
phenology <- left_join(phenology,all_spdata %>% select(group,species) %>% distinct(), by = c("scientific_name"="species"))

#dataframe of unique lcoations
unique_locs <- all_spdata %>% 
  select(species,lat,lon) %>% 
  distinct() %>% 
  left_join(filter(phenology,value), by = c("species" = "scientific_name"))

location_and_month <- unique_locs %>% select(group,lat,lon,month,month_num) %>% distinct()


pas_meta_data <- all_spdata %>% group_by(species,group) %>%
  summarise(n= n(),
            unique_locs = length(unique(paste(lat,lon)))
            )

pas_meta_data$n_months <- 
  pas_meta_data$n_species_overlap <- 
  pas_meta_data$n_possible_pas_locs <- 0



#in a loop because it was a headache working out how to do this nearly in dplyr
for (i in 1:nrow(pas_meta_data)){
  print(paste0(i/nrow(pas_meta_data)*100,"% complete"))
  species_name <- pas_meta_data$species[i]
  
  # get the months in which the target species is active
  species_months <- phenology %>% filter(scientific_name == species_name ,value) %>% pull(month)
  pas_meta_data$n_months[i] <- length(species_months)
  
  pas_meta_data$n_species_overlap[i] <- length(phenology %>% filter(month %in% species_months,group == pas_meta_data$group[i]) %>% select(scientific_name) %>% distinct() %>% pull(scientific_name))
  
  # get all the possible locations that psuedoabsences could be placed given that the species is only active in certain months
  all_locs <- location_and_month %>% filter(month %in% species_months,group == pas_meta_data$group[i]) %>% select(lat,lon) %>% distinct()
  
  # remove locations where records have already been made for the target species
  pas_locs <- setdiff(all_locs,all_spdata %>% filter(species == species_name) %>% select(lat,lon))
  
  #number of possible unique locations for pseudoabsences
  pas_meta_data$n_possible_pas_locs[i] <- nrow(pas_locs)
  
}

head(pas_meta_data)

saveRDS(pas_meta_data,"data/derived_data/species/pas_meta_data.RDS")

# pas_meta_data %>% 
#   filter(group == "butterfly") %>%
#   ggplot(aes(x = unique_locs,y = n_possible_pas_locs))+
#   geom_point()+
#   ylim(c(300000,600000))+
#   geom_smooth()



```


Generate phenologically-aware targeted pseudo-absences

```{r generate_pas}

#function for generating psuedo absences ("pas")
#loads data from environment
generate_pas <- function(sci_name){
  #get the phenology from the phenology data fraome
  #eg:
  #[1] 4 5 6 7 8 9
  species_phenology <- phenology %>% 
    filter(sci_name == scientific_name, value) %>% 
    pull(month_num)
  
  #get the species group ("butterfly" or "day_flying_moth" eor "ngiht_flying_moth" etc.)
  species_group <- all_spdata %>% 
    filter(species == sci_name) %>% 
    pull(group) %>% 
    unique()
  
  # all data recorded in the main activity season of the species AND species records of the target species from outside their main recording activity
  relevant_data <- all_spdata %>% 
    filter(group == species_group) %>%
    filter(month %in% species_phenology | species == sci_name)
  
  pas <- cpa(spdat = relevant_data,
             species = sci_name,
             nAbs = 10000,
             matchPres = T,
             recThresh = 5,
             replace = F)

  pas
}

#generate pas for a single species
test <- generate_pas("Pieris brassicae")

#generate pas for a set of species
species <- unique(all_spdata$species)

# use lapply
#basically seems to run in a loop but it works fine
pas <- lapply(as.list(species),FUN = generate_pas)

# in parallel using parLapply
clust <- makeCluster(n.cores-1)
clusterExport(clust, c("phenology","all_spdata","%>%","cpa","setdiff","filter","pull","distinct"))
pas <- parLapply(clust, as.list(species), generate_pas)

#give names to each item in list
names(pas) <- gsub(" ","_",tolower(species))


saveRDS(pas,"data/derived_data/species/pas.RDS")

```
