==> R/scripts/4_produce_outputs.Rmd <==
---
title: "Producing outputs"
author: "UK Centre for Ecology and Hydrology - DECIDE"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#set working directory to the project folder, not the R/scripts folder which it is by default because it's an Rmarkdown doc 
knitr::opts_knit$set(root.dir = '../../')
renv::settings$ignored.packages("rJava", persist = FALSE) #rJava needed for running models but not needed at this stage
```


### Load packages

```{r}
library(dplyr)
#library(raster)
library(terra)
terraOptions(memfrac = 0.1,verbose = T) # allowing terra to access less memory means it is always reading from file not from memory and seems to be much more reliable
library(tictoc)
```

### Parameters and file locations

```{r}
phenology <- readRDS("data/derived_data/species/phenology.RDS")
```

```{r}
#what group to build the seasonal decide score for
#group_to_process <- "butterfly" 
#group_to_process <- "day_flying_moth"
group_to_process <- "night_flying_moth" 

#are we accessing the SDMs from the objectstore via datalabs or within the project direction (for when running on JASMIN)
#sdm_location <- "datalabs"
#sdm_location  <- "project"
sdm_location <- "JASMIN"


# file locations
## Set up when the app initialises: firstly we need to point to where the raster files are located for each group

#this should really be set up in a better way

if(sdm_location == "datalabs"){
  
  #point to location of all the models
  data_location <- list(
      butterfly = "../../../../data-s3/DECIDE/sdm_outputs/butterfly/combined_model_outputs/PA_thinned_10000nAbs_2021_12_08",
      day_flying_moth = "../../../../data-s3/DECIDE/sdm_outputs/moth/combined_model_outputs/PA_thinned_10000nAbs/mean_score",
      night_flying_moth = "../../../../data-s3/DECIDE/sdm_outputs/nightflying_moth/combined_model_outputs/PA_thinned_10000nAbs_2021_12_08"
  )
  
  #How are the model output files named? presuming that the file names START with the species name
  file_name_template <- list(
      butterfly = list(
          mean = "_PA_thinned_10000nAbs_2021_12_08_mean_score_weightedmeanensemble.grd",
          variation = "_PA_thinned_10000nAbs_2021_12_08_mean_score_weightedvariationensemble.grd",
          auc = "_mean_score_aucOuts.csv"
      ),
      day_flying_moth = list(
          mean = "_PA_thinned_10000nAbs_weightedmeanensemble.grd",
          variation =  "_PA_thinned_10000nAbs_weightedvariationensemble.grd",
          auc = "_mean_score_aucOuts.csv"
      ),
      night_flying_moth = list(
          mean = "_PA_thinned_10000nAbs_2021_12_08_mean_score_weightedmeanensemble.grd",
          variation = "_PA_thinned_10000nAbs_2021_12_08_mean_score_weightedvariationensemble.grd",
          auc = "_mean_score_aucOuts.csv"
      )   
  )
  
} else if(sdm_location == "project") {
  #point to location of all the models
  data_location <- list(
      butterfly = "",
      day_flying_moth = "",
      night_flying_moth = ""
  )
  
  #How are the model output files named?
  file_name_template <- list(
      butterfly = list(
          mean = "",
          variation = "",
          auc = ""
      ),
      day_flying_moth = list(
          mean = "",
          variation =  "",
          auc = ""
      ),
      night_flying_moth = list(
          mean = "",
          variation = "",
          auc = ""
      )   
  )
} else if(sdm_location == "JASMIN"){
  data_location <- list(
      # butterfly = "",
      # day_flying_moth = "",
      night_flying_moth = "../../thoval/DECIDE/SDMs/outputs/nightflying_moth/combined_model_outputs/PA_thinned_10000nAbs_2021_12_08"
  )
  
  file_name_template <- list(
      # butterfly = list(
      #     mean = "",
      #     variation = "",
      #     auc = ""
      # ),
      # day_flying_moth = list(
      #     mean = "",
      #     variation =  "",
      #     auc = ""
      # ),
      night_flying_moth = list(
          mean = "_PA_thinned_10000nAbs_2021_12_08_mean_score_weightedmeanensemble.grd",
          variation = "_PA_thinned_10000nAbs_2021_12_08_mean_score_weightedvariationensemble.grd",
          auc = "_mean_score_aucOuts.csv"
      )   
  )
  
}

```

### Check files avilability

Taking the species list defined in `species_list.RDS` then building lists of files that we are looking for then comparing this to the files available.

```{r}
#get species list
species_list <- readRDS("data/derived_data/species/species_list.RDS") %>% 
  filter(group == group_to_process) %>% 
  pull(species) %>%
  #tolower() %>% 
  gsub(" ","_",.)

phenology <- phenology %>% filter(scientific_name %in% (species_list %>% gsub("_"," ",.)))


#check ensemble models
#files we want
if(sdm_location == "datalabs"){
  prediction_files <- paste0(species_list,file_name_template[[group_to_process]][["mean"]])
  sd_files <- paste0(species_list,file_name_template[[group_to_process]][["variation"]])
}

if(sdm_location == "project"){
  prediction_files <- paste0("mean_prediction_",species_list,".gri")
  sd_files <- paste0("bootstrapped_sd_",species_list,".gri")
}

file_check_df <- data.frame(species = species_list,
           prediction_file = prediction_files,
           sd_file = sd_files)

#files we have
files_available <- list.files(data_location[[group_to_process]],pattern = ".grd")


#check the target files against files that are available and add a new column to the df reporting on this check
file_check_df$prediction_check <- file_check_df$prediction_file %in% files_available
file_check_df$sd_check <- file_check_df$sd_file %in% files_available

#highlight missing species models but continues none-the-less and makes the outputs with whatever species it has predictions for.

species_to_use <- file_check_df %>% filter(prediction_check, prediction_check) %>% dplyr::select(species,prediction_check,sd_check) %>% pull(species)

print("Species with predictions:")
#species_to_use

#species with missing predictions
print("Species with missing predictions:")
file_check_df %>% filter(!prediction_check |!prediction_check) %>% dplyr::select(species,prediction_check,sd_check)

#file_check_df

#get the filenames we want to use
file_check_df <- file_check_df %>% filter(sd_check,prediction_check)
prediction_files <- file_check_df$prediction_file
sd_files <- file_check_df$sd_file
species_to_use <- file_check_df$species

#for testing, slim down the number of files
# if(F){
#   prediction_files <- prediction_files[1:40]
#   sd_files <- sd_files[1:40]
#   species_to_use <- species_to_use[1:40]
# }

```

### Load files and build rasters

Load phenology

Load raster layers for each species
When I say 'load', `terra` isn't actually loading the files into memory but creating a link. You can see this in the `mem_info()` results.

```{r}
tic("Load in species rasters from file")

#prediction raster
prediction_raster <- terra::rast(paste0(data_location[[group_to_process]],"/",prediction_files))
names(prediction_raster) <- species_to_use
prediction_raster
mem_info(prediction_raster)
raster_extent <- ext(prediction_raster[[1]])

#sd raster
sd_raster <- terra::rast(paste0(data_location[[group_to_process]],"/",sd_files))
names(sd_raster) <- species_to_use
sd_raster
mem_info(sd_raster)

toc()

```

All-time and monthly species richness and uncertainty

```{r}
#specify where the files should be saved to
richness_output_location <- paste0("data/derived_data/combined_model_outputs/",group_to_process,"/species_richness/")
uncertainty_output_location <- paste0("data/derived_data/combined_model_outputs/",group_to_process,"/model_uncertainty/")


#moths in lowercase
months_lower <- tolower(month.abb)

tic("Build all time richness")
#build all time richness and model uncertainty layers
terra::app(prediction_raster,sum,filename = paste0(richness_output_location,"raster_all_year.tif"),overwrite = T)
toc()

tic("Build all time uncertainty")
terra::app(sd_raster,mean,filename =  paste0(uncertainty_output_location,"raster_all_year.tif"),overwrite = T)
toc()

#seasonal
#looping through richness then uncertainty in the hope that accessing a consistent set of files makes it faster (this might be superstition)
for(product in c("richness","uncertainty")){
  tic(paste0("Build seasonal ",product))
  for (i in 1:12){
    tic(paste0("Build seasonal ",product," -- ",months_lower[i]))
    species <- phenology %>% filter(month_num == i,value) %>% pull(scientific_name) %>% gsub(" ","_",.)
  
    if (length(species)!=0){
      if(product == "richness"){
        prediction_raster[[names(prediction_raster) %in% species]] %>% 
          terra::app(.,sum,filename = paste0(richness_output_location,group_to_process,"_species_richness_raster_",months_lower[i],".tif"),overwrite = T)
      } else {
        sd_raster[[names(sd_raster) %in% species]] %>% 
          terra::app(.,mean, filename = paste0(uncertainty_output_location,group_to_process,"_sdm_var_raster_",months_lower[i],".tif"),overwrite = T)
      }
    }
    
    toc()
  }
  toc()
}


```

### Save rasters

Files are saved as `.tif` files in the previous code chunk. This next chunk save files as .grd/.gri files (all layers into one file)

This has been written because the app v0.4.0 and prior used .gri/.grd files whereas going forwards we'll be using .tif files. We no longer use .gri files in the app so this code is no longer essential but is kept in case we do want .gri/.grd files for whatever reason.

```{r}
#define the file names
richness_tifs <- paste0(richness_output_location,"raster_",c(months_lower,"all_year"),".tif")
uncertainty_tifs <- paste0(uncertainty_output_location,"raster_",c(months_lower,"all_year"),".tif")

#load in the files and change the names to be months
richness_stack <- rast(richness_tifs)
names(richness_stack) <- c(months_lower,"all_year")

uncertainty_stack <- rast(uncertainty_tifs)
names(uncertainty_stack) <- c(months_lower,"all_year")

#save
writeRaster(richness_stack,paste0(richness_output_location,"raster_all_months_richness.grd"),overwrite = T)
writeRaster(uncertainty_stack,paste0(uncertainty_output_location,"raster_all_months_uncertainty.grd"),overwrite = T)



```

Diagnostic plotting etc.

```{r}
#plot
plot(richness_stack)
plot(uncertainty_stack)

smaller_extent <- ext(c(455000,465000,185000,195000)) # Wallingford

richness_stack %>% crop(smaller_extent) %>% plot(range = c(0,25),col= viridisLite::viridis(n = 50) )
uncertainty_stack %>% crop(smaller_extent) %>% plot(range=c(0,0.015),col= viridisLite::viridis(n = 50))



```
